requirements.txt
numpy
pandas
qtpy
PyQt5
pyqtgraph
python-dateutil
git+https://github.com/keziah55/CustomPyQtObjects.git